Promote im2col_out from experimental; drop experimental im2col
Experimental im2col has identical kernel as existing. The only
addition is im2col_out variant. Promote it to stable, remove
experimental.

Changes:
- ops/im2col.py: add im2col_out, remove is_cuda assert
- ops/__init__.py: export im2col_out
- __init__.py: register ("im2col.out", im2col_out)
- conf/operators.yaml: add im2col_out entry
- tests/test_im2col.py: add test_im2col_out
- benchmark/test_im2col.py: add test_im2col_out
- Remove experimental_ops/im2col.py and its tests

Closes #4690
