+++ src/flag_gems/ops/__init__.py
@@ -351,7 +351,7 @@ from flag_gems.ops.i0 import i0, i0_out
 from flag_gems.ops.i0_ import i0_
 from flag_gems.ops.igamma_ import igamma_
 from flag_gems.ops.igammac_ import igammac_
-from flag_gems.ops.im2col import im2col
+from flag_gems.ops.im2col import im2col, im2col_out
 from flag_gems.ops.index import index
 from flag_gems.ops.index_add import index_add, index_add_
 from flag_gems.ops.index_copy_ import index_copy, index_copy_
@@ -1132,6 +1132,7 @@ __all__ = [
     "igamma_",
     "igammac_",
     "im2col",
+    "im2col_out",
     "index",
     "index_add",
     "index_add_",
