+++ tests/test_im2col.py
@@ -29,7 +29,6 @@ IM2COL_CONFIGS = [
 ]
 
 
-@pytest.mark.skipif(not torch.cuda.is_available(), reason="CUDA is not available")
 @pytest.mark.im2col
 @pytest.mark.parametrize("shape", IM2COL_SHAPES)
 @pytest.mark.parametrize("dtype", utils.FLOAT_DTYPES)
@@ -43,3 +42,19 @@ def test_im2col(shape, dtype, kernel_size, dilation, padding, stride):
         act_out = torch.ops.aten.im2col(x, kernel_size, dilation, padding, stride)
 
     utils.gems_assert_close(act_out, ref_out, dtype=dtype)
+
+
+@pytest.mark.im2col_out
+@pytest.mark.parametrize("shape", IM2COL_SHAPES)
+@pytest.mark.parametrize("dtype", utils.FLOAT_DTYPES)
+@pytest.mark.parametrize("kernel_size, dilation, padding, stride", IM2COL_CONFIGS)
+def test_im2col_out(shape, dtype, kernel_size, dilation, padding, stride):
+    x = torch.randn(shape, dtype=dtype, device=flag_gems.device)
+    ref_x = utils.to_reference(x)
+
+    ref_out = torch.ops.aten.im2col(ref_x, kernel_size, dilation, padding, stride)
+    out = torch.empty_like(ref_out, device=flag_gems.device)
+    with flag_gems.use_gems():
+        torch.ops.aten.im2col(x, kernel_size, dilation, padding, stride, out=out)
+
+    utils.gems_assert_close(out, ref_out, dtype=dtype)
